fonts and more cache friendly design

+++ src/Systems/DamageSystem.h
@@ -20,28 +20,18 @@ public:
     Entity entity1 = e.entity1;
     Entity entity2 = e.entity2;
 
-    Logger::Log("entity1 is player: " +
-                std::to_string(entity1.HasTag("player")));
-    Logger::Log("entity2 is player: " +
-                std::to_string(entity2.HasTag("player")));
     if (entity1.BelongsGroup("projectile") && entity2.HasTag("player")) {
       OnProjectileHitPlayer(entity1, entity2);
     } else if (entity1.HasTag("player") && entity2.BelongsGroup("projectile")) {
       OnProjectileHitPlayer(entity2, entity1);
     }
-    //
-    // if (entity1.HasTag("player") && entity2.BelongsGroup("projectile")) {
-    //   OnProjectileHitPlayer(entity2, entity1);
-    // }
-    //
-    // if (entity1.HasTag("player") && entity2.BelongsGroup("enemy")) {
-    //   // OnEnemyHitPlayer(entity1, entity2);
-    // }
-    //
-    // if (entity1.BelongsGroup("enemy") && entity2.BelongsGroup("projectile"))
-    // {
-    //   // OnEnemyHitByProjectile(entity1, entity2);
-    // }
+
+    if (entity1.BelongsGroup("enemy") && entity2.BelongsGroup("projectile")) {
+      OnProjectileHitEnemy(entity1, entity2);
+    } else if (entity1.BelongsGroup("projectile") &&
+               entity2.BelongsGroup("enemy")) {
+      OnProjectileHitEnemy(entity2, entity1);
+    }
   }
 
   void OnProjectileHitPlayer(Entity &projectile, Entity &player) {
@@ -52,6 +42,19 @@ public:
       health.currentHealth -= projectileComponent.damage;
       if (health.currentHealth <= 0) {
         player.Kill();
+      }
+      projectile.Kill();
+    }
+  }
+
+  void OnProjectileHitEnemy(Entity &enemy, Entity &projectile) {
+    auto projectileComponent = projectile.GetComponent<ProjectileComponent>();
+    if (projectileComponent.isFriendly) {
+      HealthComponent &health = enemy.GetComponent<HealthComponent>();
+
+      health.currentHealth -= projectileComponent.damage;
+      if (health.currentHealth <= 0) {
+        enemy.Kill();
       }
       projectile.Kill();
     }
